Refactor switch operation code for interface
