+++ switch_operations/cisco/nexus_interface/nexus_interface.go
@@ -1,4 +1,4 @@
-package api
+package nexus_interface
 
 type CommandRunner interface {
 	Run(string) (string, error)
